Update to 1.20.1, and make some changes including multiple creative tabs, commiting to 3d-modeled capture projectile, and botania-style versioning of this mod

+++ src/main/java/com/beesechurger/flyingfamiliars/entity/ai/FamiliarFlyingPathNavigation.java
@@ -2,9 +2,11 @@ package com.beesechurger.flyingfamiliars.entity.ai;
 
 import com.beesechurger.flyingfamiliars.entity.common.BaseFamiliarEntity;
 import net.minecraft.core.BlockPos;
+import net.minecraft.core.Position;
 import net.minecraft.core.Vec3i;
 import net.minecraft.world.entity.ai.navigation.FlyingPathNavigation;
 import net.minecraft.world.level.Level;
+import net.minecraft.world.phys.Vec3;
 
 public class FamiliarFlyingPathNavigation extends FlyingPathNavigation
 {
@@ -26,7 +28,7 @@ public class FamiliarFlyingPathNavigation extends FlyingPathNavigation
                 mob.getMoveControl().setWantedPosition(target.getX(), target.getY(), target.getZ(), 1.0f);
 
                 maxDistanceToWaypoint = mob.getBbWidth() * mob.getBbWidth();
-                Vec3i position = new Vec3i(getTempMobPos().x, getTempMobPos().y, getTempMobPos().z);
+                Vec3i position = new Vec3i((int) getTempMobPos().x, (int) getTempMobPos().y, (int) getTempMobPos().z);
 
                 if (target.distSqr(position) <= maxDistanceToWaypoint)
                     path = null;
